+++ src/Symfony/Component/ObjectMapper/Attribute/Map.php
@@ -24,12 +24,14 @@ class Map
      * @param string|class-string|null                                                                                 $target    The property or the class to map to
      * @param string|bool|callable(mixed, object): bool|null                                                           $if        A boolean, a service id or a callable that instructs whether to map
      * @param (string|callable(mixed, object, ?object): mixed)|(string|callable(mixed, object, ?object): mixed)[]|null $transform A service id or a callable that transforms the value during mapping
+     * @param array                                                                                                    $context   Extra information passed to transformers
      */
     public function __construct(
         public readonly ?string $target = null,
         public readonly ?string $source = null,
         public readonly mixed $if = null,
         public readonly mixed $transform = null,
+        public readonly array $context = [],
     ) {
     }
 }