+++ src/Symfony/Component/ObjectMapper/Tests/Fixtures/TransformCollection/TransformCollectionE.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Symfony\Component\ObjectMapper\Tests\Fixtures\TransformCollection;
+
+class TransformCollectionE
+{
+    public function __construct(
+        public string $baz,
+    ) {
+    }
+}